YCBO: make product form reusable
So it can be used for product creation as well as for product edits.
