+++ src/ya-comiste-backoffice/translations/source_strings.json
@@ -100,6 +100,20 @@
   },
   {
    "hashToText": {
+    "DG4YwlsppzoioQ32fEtyQQ==": "Something unexpected happened and server could not process the request! 🙈"
+   },
+   "filepath": "src/products/CreateProductForm.js",
+   "line_beg": 99,
+   "col_beg": 19,
+   "line_end": 102,
+   "col_end": 11,
+   "desc": "generic failure message after creating a product",
+   "project": "",
+   "type": "text",
+   "jsfbt": "Something unexpected happened and server could not process the request! 🙈"
+  },
+  {
+   "hashToText": {
     "c0PkRJ4iPzn52mQA9ajdTQ==": "At least one product name version (english or spanish) is required"
    },
    "filepath": "src/products/createProductFormSchema.js",
@@ -156,6 +170,90 @@
   },
   {
    "hashToText": {
+    "aPUJdpL8vp5qfKNsFMmsDw==": "Product name (English)"
+   },
+   "filepath": "src/products/form/ProductForm.js",
+   "line_beg": 45,
+   "col_beg": 14,
+   "line_end": 45,
+   "col_end": 98,
+   "desc": "form field name for product name in english",
+   "project": "",
+   "type": "text",
+   "jsfbt": "Product name (English)"
+  },
+  {
+   "hashToText": {
+    "If15Rd9UqwSa2ctF1W+gGw==": "Product name (Spanish)"
+   },
+   "filepath": "src/products/form/ProductForm.js",
+   "line_beg": 53,
+   "col_beg": 14,
+   "line_end": 53,
+   "col_end": 98,
+   "desc": "form field name for product name in spanish",
+   "project": "",
+   "type": "text",
+   "jsfbt": "Product name (Spanish)"
+  },
+  {
+   "hashToText": {
+    "fIkIODIEZUcYs7jnbKCMbQ==": "Product description (English)"
+   },
+   "filepath": "src/products/form/ProductForm.js",
+   "line_beg": 61,
+   "col_beg": 14,
+   "line_end": 63,
+   "col_end": 20,
+   "desc": "form field name for product description in english",
+   "project": "",
+   "type": "text",
+   "jsfbt": "Product description (English)"
+  },
+  {
+   "hashToText": {
+    "0JhgWnXU3GS2BSJv0aKTSA==": "Product description (Spanish)"
+   },
+   "filepath": "src/products/form/ProductForm.js",
+   "line_beg": 71,
+   "col_beg": 14,
+   "line_end": 73,
+   "col_end": 20,
+   "desc": "form field name for product description in spanish",
+   "project": "",
+   "type": "text",
+   "jsfbt": "Product description (Spanish)"
+  },
+  {
+   "hashToText": {
+    "r84jIlbvbpbZx8MqJ0A94w==": "Price (MXN)"
+   },
+   "filepath": "src/products/form/ProductForm.js",
+   "line_beg": 81,
+   "col_beg": 14,
+   "line_end": 81,
+   "col_end": 77,
+   "desc": "form field name for product price",
+   "project": "",
+   "type": "text",
+   "jsfbt": "Price (MXN)"
+  },
+  {
+   "hashToText": {
+    "lZD4GtdtmFDnxD98aPVEUw==": "Product pictures"
+   },
+   "filepath": "src/products/form/ProductForm.js",
+   "line_beg": 89,
+   "col_beg": 14,
+   "line_end": 89,
+   "col_end": 85,
+   "desc": "form field name for product pictures",
+   "project": "",
+   "type": "text",
+   "jsfbt": "Product pictures"
+  },
+  {
+   "hashToText": {
     "0BsTiPapD1k7SKNIyzq9Og==": "Products inventory"
    },
    "filepath": "pages/products.js",
@@ -184,101 +282,59 @@
   },
   {
    "hashToText": {
-    "DG4YwlsppzoioQ32fEtyQQ==": "Something unexpected happened and server could not process the request! 🙈"
+    "D+WsTLSMdok9SuNN16agEQ==": "Create a new product"
    },
    "filepath": "pages/products/create.js",
-   "line_beg": 97,
-   "col_beg": 19,
-   "line_end": 100,
-   "col_end": 11,
-   "desc": "generic failure message after creating a product",
+   "line_beg": 15,
+   "col_beg": 10,
+   "line_end": 15,
+   "col_end": 82,
+   "desc": "create a new product page heading",
    "project": "",
    "type": "text",
-   "jsfbt": "Something unexpected happened and server could not process the request! 🙈"
+   "jsfbt": "Create a new product"
   },
   {
    "hashToText": {
-    "aPUJdpL8vp5qfKNsFMmsDw==": "Product name (English)"
+    "kZSl2UcMZDzRSCGJm9hDkw==": "Products inventory"
    },
    "filepath": "pages/products/create.js",
-   "line_beg": 118,
-   "col_beg": 16,
-   "line_end": 118,
-   "col_end": 100,
-   "desc": "form field name for product name in english",
+   "line_beg": 21,
+   "col_beg": 17,
+   "line_end": 21,
+   "col_end": 91,
+   "desc": "go back to products navigation button",
    "project": "",
    "type": "text",
-   "jsfbt": "Product name (English)"
+   "jsfbt": "Products inventory"
   },
   {
    "hashToText": {
-    "If15Rd9UqwSa2ctF1W+gGw==": "Product name (Spanish)"
+    "kZSl2UcMZDzRSCGJm9hDkw==": "Products inventory"
    },
-   "filepath": "pages/products/create.js",
-   "line_beg": 126,
-   "col_beg": 16,
-   "line_end": 126,
-   "col_end": 100,
-   "desc": "form field name for product name in spanish",
+   "filepath": "pages/products/edit/[productKey].js",
+   "line_beg": 28,
+   "col_beg": 17,
+   "line_end": 28,
+   "col_end": 91,
+   "desc": "go back to products navigation button",
    "project": "",
    "type": "text",
-   "jsfbt": "Product name (Spanish)"
+   "jsfbt": "Products inventory"
   },
   {
    "hashToText": {
-    "fIkIODIEZUcYs7jnbKCMbQ==": "Product description (English)"
+    "wx008fc5s2Mij2EDAuzNIw==": "Delete product (TODO)"
    },
-   "filepath": "pages/products/create.js",
-   "line_beg": 134,
-   "col_beg": 16,
-   "line_end": 136,
-   "col_end": 22,
-   "desc": "form field name for product description in english",
+   "filepath": "pages/products/edit/[productKey].js",
+   "line_beg": 32,
+   "col_beg": 17,
+   "line_end": 32,
+   "col_end": 89,
+   "desc": "delete product navigation button",
    "project": "",
    "type": "text",
-   "jsfbt": "Product description (English)"
-  },
-  {
-   "hashToText": {
-    "0JhgWnXU3GS2BSJv0aKTSA==": "Product description (Spanish)"
-   },
-   "filepath": "pages/products/create.js",
-   "line_beg": 144,
-   "col_beg": 16,
-   "line_end": 146,
-   "col_end": 22,
-   "desc": "form field name for product description in spanish",
-   "project": "",
-   "type": "text",
-   "jsfbt": "Product description (Spanish)"
-  },
-  {
-   "hashToText": {
-    "r84jIlbvbpbZx8MqJ0A94w==": "Price (MXN)"
-   },
-   "filepath": "pages/products/create.js",
-   "line_beg": 154,
-   "col_beg": 16,
-   "line_end": 154,
-   "col_end": 79,
-   "desc": "form field name for product price",
-   "project": "",
-   "type": "text",
-   "jsfbt": "Price (MXN)"
-  },
-  {
-   "hashToText": {
-    "lZD4GtdtmFDnxD98aPVEUw==": "Product pictures"
-   },
-   "filepath": "pages/products/create.js",
-   "line_beg": 162,
-   "col_beg": 16,
-   "line_end": 162,
-   "col_end": 87,
-   "desc": "form field name for product pictures",
-   "project": "",
-   "type": "text",
-   "jsfbt": "Product pictures"
+   "jsfbt": "Delete product (TODO)"
   },
   {
    "hashToText": {
