+++ src/ya-comiste-backoffice/translations/src_manifest.json
@@ -1,1 +1,1 @@
-{"translations/enum_manifest.json":["src/AuthButtons.js","src/Navigation.js","src/products/createProductFormSchema.js","pages/products.js","pages/products/create.js","pages/users.js"]}
+{"translations/enum_manifest.json":["src/AuthButtons.js","src/Navigation.js","src/products/CreateProductForm.js","src/products/createProductFormSchema.js","src/products/form/ProductForm.js","pages/products.js","pages/products/create.js","pages/products/edit/[productKey].js","pages/users.js"]}
